requirements_dev.txt
attrs>=19<20
cached-property>=1<2
Click>=7.0<8.0
emcee>=3.0rc2<4.0
GetDist>=0.3<1.0
matplotlib>=3<4
pytest>=4<5
PyYAML>=5.1<6
scipy>=1.2<2
Sphinx>=2.0<3
pre-commit
tox
flake8
flake8-docstrings
flake8-black
flake8-builtins
flake8-logging-format
flake8-rst-docstrings
flake8-rst
flake8-markdown
flake8-bugbear
flake8-comprehensions
flake8-print
cohesion
pydocstyle<4.0
